fixed loading gif and lightbox
