+++ client/package.json
@@ -9,6 +9,7 @@
     "react": "^16.8.2",
     "react-apollo": "^2.4.1",
     "react-dom": "^16.8.2",
+    "react-image-lightbox": "^5.1.0",
     "react-linkify": "^0.2.2",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.5",
