+++ client/src/components/GameItems.js
@@ -1,43 +1,8 @@
 import React from "react";
-import Lightbox from 'react-image-lightbox';
-import 'react-image-lightbox/style.css';
 import { ButtonDropdown, DropdownToggle, DropdownMenu, DropdownItem } from 'reactstrap';
 import { Mutation } from "react-apollo";
 import gql from "graphql-tag";
-
-class GameLightbox extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      photoIndex: 0,
-      isOpen: false,
-    };
-  }
-
-  render() {
-    const { isOpen } = this.state;
-    const { img } = this.props;
-    // const { imgUrl } = this.props.single.node;
-
-    return (
-      <div>
-        <img 
-          src={img} 
-          width={'100px'} 
-          style={{ marginTop: 15 }} 
-          type="button" onClick={() => this.setState({ isOpen: true })}
-        />
-        {isOpen && (
-          <Lightbox
-            mainSrc={img}
-            onCloseRequest={() => this.setState({ isOpen: false })}
-          />
-        )}
-      </div>
-    );
-  }
-}
+import GameLightbox from './GameLightbox';
 
 const ADD_GAME_TO_USER = gql`
     mutation AddGameToCollection($input: Gamecatalog_updateGameInput!) {
