+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "start": "node server.js",
     "server": "nodemon server.js",
-    "client": "npm start --prefix client",
-    "dev": "concurrently \"npm run server\" \"npm run client\""
+    "client": "cd client && yarn start",
+    "dev": "concurrently \"yarn run server\" \"yarn run client\""
   },
   "author": "William Cruz",
   "license": "MIT",
@@ -17,8 +17,7 @@
     "cors": "^2.8.5",
     "express": "^4.16.4",
     "express-graphql": "^0.7.1",
-    "graphql": "^14.1.1",
-    "react-image-lightbox": "^5.1.0"
+    "graphql": "^14.1.1"
   },
   "devDependencies": {
     "nodemon": "^1.18.10"
